Rest of the created files
Some of this files were created by default when we use the command for the devise model, and some
of them like the development file and application file were modified only one line because of the configuration of devise.
The schema file was modified when we create the model User where was added the table users.

Also we create a migration to delete an useless table

Co-authored-by: RobertoCarMed <betocarrasco3652@gmail.com>